+++ services/scheduler-service/src/controllers/calendar.controller.ts
@@ -97,7 +97,7 @@ export class CalendarController {
   @authorize([PermissionKey.CreateCalendar])
   @post('/calendars/calendarSubscription', {
     responses: {
-      '200': {
+      [STATUS_CODE.OK]: {
         description: calendarModelInstance,
         content: {'application/json': {schema: getModelSchemaRef(CalendarDTO)}},
       },
@@ -123,26 +123,36 @@ export class CalendarController {
 
     delete calendarDTO.subscription;
     let response = await this.calendarService.createCalendar(calendarDTO);
-
     let identifierType = this.schdulerConfig?.identifierMappedTo;
     if (!identifierType) {
       identifierType = IdentifierType.Id;
     }
 
-    const subscriptionList = await this.subscriptionRepository.find({
-      where: {
-        identifier: this.currentUser[identifierType],
-      },
-    });
+    let subscriptionIdentifier;
+    if (subscription.identifier) {
+      subscriptionIdentifier = subscription.identifier;
+    } else {
+      subscriptionIdentifier = this.currentUser[identifierType];
+    }
 
+    const where = {
+      and: [{identifier: subscriptionIdentifier}, {isPrimary: true}],
+    };
+
+    const subscriptionList = await this.subscriptionRepository.find({where});
     if (subscriptionList.length > 0) {
       subscription.isPrimary = false;
     } else {
       subscription.isPrimary = true;
     }
+
     if (response.id) {
       subscription.calendarId = response.id;
-      subscription.identifier = calendarDTO.identifier;
+      if (subscriptionIdentifier) {
+        subscription.identifier = subscriptionIdentifier;
+      } else {
+        throw new HttpErrors.NotFound(ErrorKeys.SubscriptionIdentifierNotExist);
+      }
       subscription.accessRole = AccessRoleType.Owner;
       const subscriptionResponse = await this.subscriptionRepository.create(
         subscription,