+++ services/scheduler-service/src/services/event.service.ts
@@ -3,9 +3,9 @@ import {repository, Filter} from '@loopback/repository';
 import {ResponseStatusType} from '../models/enums/response-status.enum';
 import {IStartEndTime, EventAttendeeView} from '../models';
 import {EventAttendeeViewRepository} from '../repositories/event-attendee-view.repository';
-import {EventAttendeeViewItemDTO} from '../models/EventAttendeeViewItemDTO.dto';
+import {EventAttendeeViewItemDTO} from '../models/event-attendee-view-item.dto';
 import {HttpErrors} from '@loopback/rest';
-import {ErrorKeys} from '../models/enums';
+import {ErrorKeys, StatusType} from '../models/enums';
 
 @bind({scope: BindingScope.TRANSIENT})
 export class EventService {
@@ -20,6 +20,9 @@ export class EventService {
     timeMin: Date,
   ) {
     const where = [];
+    where.push({
+      status: {neq: StatusType.Cancelled},
+    });
     where.push({
       or: [
         {responseStatus: {neq: ResponseStatusType.Declined}},